rewrote Makefile
Made the targets all, install, uninstall and clean phony to avoid interference with certain filesystems on certain systems (basically, the code would never even run if a file called clean or install is present which may very well be the case if some distro decides to include this in their repo). Also clean shouldn't depend on all because that will cause the file to be built only to be deleted immediately afterwards which wastes precious CPU cycles. Also, I updated the build rule itself to make it more generic